+++ build-tools/docker-create.sh
@@ -3,7 +3,7 @@
 # Delete all containers
 # $ docker rm -f $(docker ps -a -q)
 # 
-# Delete all volumns
+# Delete all volumes
 # $ docker volume rm $(docker volume ls -q)
 # 
 # Delete all images