+++ sea-orm-macros/src/lib.rs
@@ -639,6 +639,46 @@ pub fn derive_relation(input: TokenStream) -> TokenStream {
         .into()
 }
 
+/// The DeriveRelatedEntity derive macro will implement seaography::RelationBuilder for RelatedEntity enumeration.
+///
+/// ### Usage
+///
+/// ```ignore
+/// use sea_orm::entity::prelude::*;
+///
+/// // ...
+/// // Model, Relation enum, etc.
+/// // ...
+///
+/// #[derive(Copy, Clone, Debug, EnumIter, DeriveRelatedEntity)]
+/// pub enum RelatedEntity {
+///     #[sea_orm(entity = "super::address::Entity")]
+///     Address,
+///     #[sea_orm(entity = "super::payment::Entity")]
+///     Payment,
+///     #[sea_orm(entity = "super::rental::Entity")]
+///     Rental,
+///     #[sea_orm(entity = "Entity", def = "Relation::SelfRef.def()")]
+///     SelfRef,
+///     #[sea_orm(entity = "super::store::Entity")]
+///     Store,
+///     #[sea_orm(entity = "Entity", def = "Relation::SelfRef.def().rev()")]
+///     SelfRefRev,
+/// }
+/// ```
+#[cfg(feature = "derive")]
+#[proc_macro_derive(DeriveRelatedEntity, attributes(sea_orm))]
+pub fn derive_related_entity(input: TokenStream) -> TokenStream {
+    let input = parse_macro_input!(input as DeriveInput);
+    if cfg!(feature = "seaography") {
+        derives::expand_derive_related_entity(input)
+            .unwrap_or_else(Error::into_compile_error)
+            .into()
+    } else {
+        TokenStream::new()
+    }
+}
+
 /// The DeriveMigrationName derive macro will implement `sea_orm_migration::MigrationName` for a migration.
 ///
 /// ### Usage