+++ tests/common/mod.rs
@@ -13,6 +13,9 @@ pub struct TestContext {
 
 impl TestContext {
     pub async fn new(test_name: &str) -> Self {
+        dotenv::from_filename(".env.local").ok();
+        dotenv::from_filename(".env").ok();
+
         let base_url =
             std::env::var("DATABASE_URL").expect("Enviroment variable 'DATABASE_URL' not set");
         let db: DatabaseConnection = setup::setup(&base_url, test_name).await;